requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
flask-login>=0.6.3
flask-wtf>=1.2.0
gunicorn>=21.0.0
huggingface-hub>=0.23.0
python-dotenv>=1.0.0
pypdf>=4.0.0
werkzeug>=3.0.0
